Moved HashicorpNode into testFixtures
