+++ acceptance-tests/build.gradle
@@ -24,7 +24,6 @@ dependencies {
   testImplementation 'org.junit.jupiter:junit-jupiter-api'
   testImplementation 'io.vertx:vertx-core'
   testImplementation 'io.vertx:vertx-web'
-  testImplementation 'org.apache.tuweni:tuweni-net'
 
   testImplementation 'org.apache.logging.log4j:log4j-api'
   testImplementation 'org.assertj:assertj-core'
