+++ keystorage/hashicorp/build.gradle
@@ -57,6 +57,16 @@ dependencies {
   integrationTestImplementation 'org.mock-server:mockserver-netty'
   integrationTestImplementation sourceSets.testFixtures.output
 
+  testFixturesImplementation 'com.github.docker-java:docker-java'
+  testFixturesImplementation 'org.apache.logging.log4j:log4j-core'
+  testFixturesImplementation 'org.apache.logging.log4j:log4j-slf4j-impl'
+  testFixturesImplementation 'org.apache.logging.log4j:log4j-api'
+  testFixturesImplementation 'org.assertj:assertj-core'
+  testFixturesImplementation 'org.awaitility:awaitility'
+  testFixturesImplementation 'org.apache.tuweni:tuweni-net'
+  testFixturesImplementation 'org.junit.jupiter:junit-jupiter-engine'
+  testFixturesImplementation 'org.bouncycastle:bcpkix-jdk15on'
+
 
   errorprone("com.google.errorprone:error_prone_core")
 }
