+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signing/hashicorp/HashicorpConfigUtil.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.plus.plugin.encryptedstorage.encryption.util;
+package tech.pegasys.signing.hashicorp;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.nio.file.Files.createTempFile;
