+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signing/hashicorp/dsl/hashicorp/HashicorpVaultDocker.java
@@ -10,11 +10,17 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.keystorage.hashicorp.dsl.hashicorp;
+package tech.pegasys.signing.hashicorp.dsl.hashicorp;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.containsString;
+
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.net.URI;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -37,21 +43,17 @@ import com.github.dockerjava.core.DefaultDockerClientConfig;
 import com.github.dockerjava.core.command.ExecStartResultCallback;
 import com.github.dockerjava.core.command.PullImageResultCallback;
 import com.github.dockerjava.core.command.WaitContainerResultCallback;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.net.URI;
-import java.nio.file.Path;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.assertj.core.api.Assertions;
 import org.awaitility.Awaitility;
 import org.awaitility.core.ThrowingRunnable;
+import org.hamcrest.Matchers;
 
 public class HashicorpVaultDocker {
 
-  public static String SECRET_CONTENT = "8f2a55949038a9610f50fb23b5883af3b4ecb3c3bb792cbcefbd1542c692be63";
+  public static String SECRET_CONTENT =
+      "8f2a55949038a9610f50fb23b5883af3b4ecb3c3bb792cbcefbd1542c692be63";
 
   private static final Logger LOG = LogManager.getLogger();
   private static final String HASHICORP_VAULT_IMAGE = "vault:1.2.3";
@@ -62,7 +64,8 @@ public class HashicorpVaultDocker {
   private static final String SECRET_PATH = "/secret";
   private static final String SIGNING_KEY_RESOURCE = "/ethsignerSigningKey";
   private static final String VAULT_PUT_RESOURCE = SECRET_PATH + SIGNING_KEY_RESOURCE;
-  private static final String VAULT_SIGNING_KEY_PATH = "/v1" + SECRET_PATH + "/data" + SIGNING_KEY_RESOURCE;
+  private static final String VAULT_SIGNING_KEY_PATH =
+      "/v1" + SECRET_PATH + "/data" + SIGNING_KEY_RESOURCE;
 
   private static final String EXPECTED_FOR_SECRET_CREATION = "created_time";
   private static final String EXPECTED_FOR_STATUS = "Sealed";
@@ -150,7 +153,7 @@ public class HashicorpVaultDocker {
         () -> {
           final ExecCreateCmdResponse execCreateCmdResponse =
               getExecCreateCmdResponse(vaultStatusCommand());
-          assertThat(
+          Assertions.assertThat(
                   runCommandInVaultContainerAndCompareOutput(
                       execCreateCmdResponse, EXPECTED_FOR_STATUS))
               .isTrue();
@@ -174,7 +177,7 @@ public class HashicorpVaultDocker {
         () -> {
           final ExecCreateCmdResponse execCreateCmdResponse =
               getExecCreateCmdResponse(vaultPutSecretCommand());
-          assertThat(
+          Assertions.assertThat(
                   runCommandInVaultContainerAndCompareOutput(
                       execCreateCmdResponse, EXPECTED_FOR_SECRET_CREATION))
               .isTrue();
@@ -192,7 +195,7 @@ public class HashicorpVaultDocker {
             .ignoreExceptions()
             .until(
                 () -> runCommandInContainerAndGetOutput(execCreateCmdResponse),
-                containsString("root_token"));
+                Matchers.containsString("root_token"));
 
     final String rootToken;
     final String unsealKey;
@@ -200,13 +203,13 @@ public class HashicorpVaultDocker {
       final JsonNode jsonNode = objectMapper.readTree(jsonOutput);
       rootToken = jsonNode.get("root_token").asText();
       final JsonNode unsealKeyJsonNodeArray = jsonNode.get("unseal_keys_b64");
-      assertThat(unsealKeyJsonNodeArray.isArray()).isTrue();
+      Assertions.assertThat(unsealKeyJsonNodeArray.isArray()).isTrue();
       unsealKey = unsealKeyJsonNodeArray.get(0).asText();
     } catch (IOException e) {
       throw new RuntimeException("Error in parsing json output from vault unseal command", e);
     }
-    assertThat(rootToken).isNotNull();
-    assertThat(unsealKey).isNotNull();
+    Assertions.assertThat(rootToken).isNotNull();
+    Assertions.assertThat(unsealKey).isNotNull();
     LOG.info("Hashicorp vault is initialized");
 
     return new HashicorpVaultTokens(unsealKey, rootToken);
@@ -228,11 +231,11 @@ public class HashicorpVaultDocker {
             .ignoreExceptions()
             .until(
                 () -> runCommandInContainerAndGetOutput(execCreateCmdResponse),
-                containsString("sealed"));
+                Matchers.containsString("sealed"));
 
     try {
       final JsonNode jsonNode = objectMapper.readTree(jsonOutput);
-      assertThat(jsonNode.get("sealed").asBoolean()).isFalse();
+      Assertions.assertThat(jsonNode.get("sealed").asBoolean()).isFalse();
     } catch (IOException e) {
       throw new RuntimeException("Error in parsing json output from vault unseal command", e);
     }
@@ -462,8 +465,8 @@ public class HashicorpVaultDocker {
 
   private int portSpec(final Ports ports) {
     final Binding[] tcpPorts = ports.getBindings().get(ExposedPort.tcp(DEFAULT_VAULT_PORT));
-    assertThat(tcpPorts).isNotEmpty();
-    assertThat(tcpPorts.length).isEqualTo(1);
+    Assertions.assertThat(tcpPorts).isNotEmpty();
+    Assertions.assertThat(tcpPorts.length).isEqualTo(1);
 
     return Integer.parseInt(tcpPorts[0].getHostPortSpec());
   }
