+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signing/hashicorp/dsl/hashicorp/HashicorpVaultTokens.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.signers.keystorage.hashicorp.dsl.hashicorp;
+package tech.pegasys.signing.hashicorp.dsl.hashicorp;
 
 public class HashicorpVaultTokens {
   private final String unsealKey;
